Refactor arquitecture and some filenames
