+++ public/controllers/register-agent/core/register-commands/README.md
