+++ public/controllers/register-agent/core/register-commands/command-generator/command-generator.ts
