+++ public/controllers/register-agent/core/register-commands/exceptions/index.ts
