+++ public/controllers/register-agent/core/register-commands/services/get-install-command.service.ts
