+++ public/controllers/register-agent/core/register-commands/services/search-os-definitions.service.test.ts
